# x86_64: align loop headers to 64 bytes