@cfallin 2022-10-04
@pepyakin , thanks for this change. However, I'm not sure that we want to take it as-is; or at least, I'd like to see more data.

The Sightglass runs you provide are all over the place -- in some cases this is better by 5-10%, in other cases `main` is better by 5-10%. There doesn't appear to be a clear trend to me.

What's more concerning, the `instantiation` times have wild swings as well, but this shouldn't have been affected at all by this change. This makes me concerned about the reliability of the measurement setup and confidence-interval computation.

Would you be willing to take a few of the larger benchmarks (`spidermonkey` and `bz2`, say) and run them by precompiling to `.cwasm`s then measuring time with your tool of choice (I like `hyperfine`) for a `wasmtime run ...` command? I'm curious how much variance this sees and what trends we'll get.

I share @bjorn3's concern about code size here as well. Aligning *every* loop to a 64-byte granularity is likely to bloat some cases nontrivially. Could you report the effect on code-size (the `.text` segment of the `.cwasm`s) for benchmarks as well?