@pepyakin 2022-10-11
That's a good catch. I must admit that I did not notice that it was about instantiation! Probably something was running on my test machine and ruined the measurement. I also agree with the points made by @bjorn3. In retrospect, that seems obvious 🤦 

I had an impression that this PR would be a quick one, but it seems I was wrong = ) I will try to further it, but that won't be my primary focus, so it will likely take some time[^1].

[^1]: Should anyone feel the acute need/or just want to take over for other reasons, feel free to take it on.