@lpereira 2024-05-21
Doing some necromancing here, but I found a nice article detailing [how the .NET JIT is padding their loops](https://kunalspathak.github.io/2021-03-15-Loop-Alignment/) and adopting a similar strategy might be useful.   It's quite a bit more involved than this PR.