@lpereira 2024-05-23
> @lpereira how are the loop / basic-block weights computed in the .NET JIT? Are they based on profiling counters or some ahead-of-time estimate? 

I don't know!

What I know, however, is that CoreCLR has a tiered JIT: AOT compilation is possible and usually performed to speed up startup, but the compiler in AOT mode does as much work optimizing code as Winch, IIRC.  Once it detects that a function needs a second look (and I don't know how it does, maybe a per-function counter and some tiny code in the epilogue that decrements and calls into a "jit this harder" function that does that, then patches the caller to become a trampoline? I really don't know; it's probably something a whole lot more clever than that), it then spends more time recompiling things.  This is pretty complicated to implement, and wouldn't help with the pure AOT scenarios we have, of course (unless we had some kind of PGO thing).

I'll spend some time looking through CoreCLR later to figure out exactly what they do and how they do it.  I'll look at how GCC and Clang does it too.

> 
> The heuristics seem to be the crux of things here: we can't align all loops because it would lead to (presumably) a significant increase in code size (how much? someone needs to do the experiments). 

Right.  I've given this some thought, and we could experiment with some heuristics, based on:

- The number of (WASM) instructions in the loop. Loops with over, I don't know, 100 instructions in the body is likely to not be that tight.
- The kind of Cranelift IR instructions in the loop. For instance:
     - If it's mostly arithmetic, SIMD, and memory read/writes, then it's probably a hot loop
     - If there are non-math-related libcalls, then it's less likely it's a hot loop
     - If there are host calls, it's also less likely it's a hot loop
     - If most calls are to small, trivial functions (say ~16 WASM instructions, no loops, maybe at most 2 branches, no other calls, or some other heuristic similar we'd use for inlining a function)
     - Maybe factor in a ratio of branches/number of instructions too, and if it's not branch-heavy, it's probably a hot loop?
- If we ever have some sort of PGO, feed that into the "loop hotness" score

It's all a bit fiddly, of course, and we'd need to empirically adjust these scores, but I think it's a doable strategy to identify hot loops.

Then, when lowering, at least on x86, two things need to happen if a loop is considered hot:

- Align the first instruction of the loop on a 32-byte boundary if there's SIMD, and 16-byte otherwise
- Ensure that no instructions in the loop crosses a cache line boundary