@alexcrichton 2025-05-27
I'm going to close this as pretty old at this point. I think it's probably best to continue discussion on https://github.com/bytecodealliance/wasmtime/issues/4883 about how best to apply this optimization